refactor: streamline user session handling in AppHeader and update admin checks
